Refactoring to attempt to issue a pre-flight query to collect whether we even need to process any further
Attempting to use this on the guardrail repo highlighted a pretty big error in my approach here, what may be more useful is to refactor the whole core to permit a single runtime and dataset to be evaluated against a number of configs during a single run. That would permit the reuse of the fetched GraphQL results across all runs, significantly reducing the cost of regenerating release drafts.